# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CargoLink Backend
  description: Smart Freight Matching & Tracking System — backend API. This is the live-generated source
    of truth for the API contract; the committed, versioned export lives at shared/openapi/openapi.yaml
    and is checked for drift by backend/tests/test_openapi_contract.py.
  version: 0.1.0
paths:
  /auth/otp/request:
    post:
      tags:
      - auth
      summary: Request an OTP for a phone number
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpRequestResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: &id001
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequestRequest'
  /auth/otp/verify:
    post:
      tags:
      - auth
      summary: Verify an OTP and receive a token pair
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpVerifyResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
  /auth/refresh:
    post:
      tags:
      - auth
      summary: Exchange a refresh token for a new access token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
...
  /loads:
    post:
      tags:
      - loads
      summary: Post a new cargo load
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadCreateResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadCreateRequest'
  /loads/{load_id}/accept:
    post:
      tags:
      - loads
      summary: Transporter accepts a matched load
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadAcceptResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: Load being accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadAcceptRequest'
...
  /documents/{load_id}:
    get:
      tags:
      - documents
      summary: Get the document checklist for a load
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentChecklistResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: Load to fetch the checklist for.
  /documents/{load_id}/upload:
    post:
      tags:
      - documents
      summary: Upload a shipment document (multipart)
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: Load this document belongs to.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
...
  /containers/{load_id}:
    post:
      tags:
      - containers
      summary: Attach container tracking data to a load
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: Load this container data belongs to.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContainerCreateRequest'
    get:
      tags:
      - containers
      summary: Read a load's container data (null for road-only loads)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: Load to fetch container data for.
  /containers/{container_id}:
    patch:
      tags:
      - containers
      summary: Update container tracking data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: container_id
        in: path
        required: true
        schema: *id002
        description: Container record to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContainerUpdateRequest'
...
components:
  schemas:
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. 'not_found', 'validation_error', 'rate_limited').
        message:
          type: string
          description: Human-readable explanation of what went wrong, safe to display to an end user.
      required:
      - code
      - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      description: Standard error envelope returned by every endpoint in this API.
    ValidationError:
      type: object
      properties:
        loc:
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          type: string
        type:
          type: string
      required:
      - loc
      - msg
      - type
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    Role:
      type: string
      enum:
      - shipper
      - transporter
      - admin
      description: The user's platform role.
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier.
        role:
          $ref: '#/components/schemas/Role'
        name:
          type:
          - string
          - 'null'
          description: Display name, if set during onboarding.
        phone:
          type: string
          pattern: ^\+[1-9]\d{6,14}$
          description: E.164 phone number used for OTP login.
        email:
          type:
          - string
          - 'null'
          format: email
          description: Optional email address.
        created_at:
          type: string
          format: date-time
          description: When this user account was created.
      required:
      - id
      - role
      - phone
      - created_at
    OtpRequestRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: ^\+[1-9]\d{6,14}$
          description: Phone number to send the OTP to, E.164 format (e.g. +919876543210).
          examples:
          - '+919876543210'
      required:
      - phone
    OtpRequestResponse:
      type: object
      properties:
        otp_sent:
          type: boolean
          description: True if the OTP was dispatched (or queued) successfully.
      required:
      - otp_sent
    OtpVerifyRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: ^\+[1-9]\d{6,14}$
          description: Phone number the OTP was requested for.
        otp:
          type: string
          pattern: ^\d{6}$
          description: The 6-digit one-time password received by SMS.
      required:
      - phone
      - otp
    OtpVerifyResponse:
      type: object
      properties:
        token:
          type: string
          description: Short-lived JWT access token (~15 min).
        refresh_token:
          type: string
          description: Longer-lived, rotating refresh token.
        user:
          $ref: '#/components/schemas/UserSummary'
      required:
      - token
      - refresh_token
      - user
    TokenRefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: A previously issued, still-valid (unrotated) refresh token.
      required:
      - refresh_token
    TokenRefreshResponse:
      type: object
      properties:
        token:
          type: string
          description: A newly issued short-lived JWT access token.
      required:
      - token
...
    GeoPoint:
      type: object
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
          description: Latitude in decimal degrees (WGS84).
        lng:
          type: number
          minimum: -180
          maximum: 180
          description: Longitude in decimal degrees (WGS84).
      required:
      - lat
      - lng
...
    CargoType:
      type: string
      enum:
      - general
      - fragile
      - hazardous
      - refrigerated
    LoadStatus:
      type: string
      enum:
      - open
      - matched
      - accepted
      - in_transit
      - delivered
      - cancelled
    LoadCreateRequest:
      type: object
      properties:
        weight:
          type: number
          exclusiveMinimum: 0
          description: Cargo weight in kilograms.
        cargo_type:
          $ref: '#/components/schemas/CargoType'
        source:
          $ref: '#/components/schemas/GeoPoint'
        destination:
          $ref: '#/components/schemas/GeoPoint'
        deadline:
          type: string
          format: date-time
          description: Latest acceptable pickup deadline.
      required:
      - weight
      - cargo_type
      - source
      - destination
      - deadline
    LoadCreateResponse:
      type: object
      properties:
        load_id:
          type: string
          format: uuid
          description: Unique identifier for the newly posted load.
        status:
          $ref: '#/components/schemas/LoadStatus'
      required:
      - load_id
      - status
    LoadAcceptRequest:
      type: object
      properties:
        vehicle_id:
          type: string
          format: uuid
          description: The vehicle being assigned to fulfil this load.
      required:
      - vehicle_id
    LoadAcceptResponse:
      type: object
      properties:
        match_id:
          type: string
          format: uuid
          description: Identifier of the created load_matches record.
        status:
          $ref: '#/components/schemas/LoadStatus'
      required:
      - match_id
      - status
...
    DocType:
      type: string
      enum:
      - commercial_invoice
      - packing_list
      - bill_of_lading_or_airway_bill
      - customs_clearance_certificate
      - certificate_of_origin
    DocumentStatus:
      type: string
      enum:
      - Pending
      - Uploaded
      - Verified
      - Cleared
    ShipmentDocumentItem:
      type: object
      properties:
        doc_type:
          $ref: '#/components/schemas/DocType'
        status:
          $ref: '#/components/schemas/DocumentStatus'
        file_url:
          type:
          - string
          - 'null'
          description: Short-lived signed URL to read the uploaded file, if any.
      required:
      - doc_type
      - status
    DocumentChecklistResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/ShipmentDocumentItem'
          description: The cargo-type-conditional checklist for this load (see Cluster F.1's rules.py).
      required:
      - documents
    DocumentUploadRequest:
      type: object
      properties:
        doc_type:
          $ref: '#/components/schemas/DocType'
      required:
      - doc_type
      description: Metadata accompanying a multipart document upload; the binary file part is handled
        via FastAPI's UploadFile, not modeled here.
    DocumentUploadResponse:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
          description: Identifier of the created/updated document record.
        status:
          $ref: '#/components/schemas/DocumentStatus'
      required:
      - doc_id
      - status
...
    ContainerCreateRequest:
      type: object
      properties:
        container_no:
          type: string
          pattern: ^[A-Z]{4}\d{7}$
          description: ISO 6346 container number, e.g. MSCU1234567.
        vessel_or_flight:
          type: string
          minLength: 1
          maxLength: 100
          description: Vessel name or flight number carrying this container.
        port_of_loading:
          type: string
          minLength: 1
          maxLength: 100
          description: Port/airport where the container was loaded.
        port_of_discharge:
          type: string
          minLength: 1
          maxLength: 100
          description: Port/airport where the container will be discharged.
      required:
      - container_no
      - vessel_or_flight
      - port_of_loading
      - port_of_discharge
    ContainerUpdateRequest:
      type: object
      properties:
        container_no:
          type:
          - string
          - 'null'
          pattern: ^[A-Z]{4}\d{7}$
          description: ISO 6346 container number.
        vessel_or_flight:
          type:
          - string
          - 'null'
          maxLength: 100
          description: Vessel name or flight number.
        port_of_loading:
          type:
          - string
          - 'null'
          maxLength: 100
          description: Port/airport where the container was loaded.
        port_of_discharge:
          type:
          - string
          - 'null'
          maxLength: 100
          description: Port/airport where the container will be discharged.
    ContainerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique container record identifier.
        load_id:
          type: string
          format: uuid
          description: The load this container data belongs to.
        container_no:
          type: string
          description: ISO 6346 container number.
        vessel_or_flight:
          type: string
          description: Vessel name or flight number.
        port_of_loading:
          type: string
          description: Port/airport where the container was loaded.
        port_of_discharge:
          type: string
          description: Port/airport where the container will be discharged.
      required:
      - id
      - load_id
      - container_no
      - vessel_or_flight
      - port_of_loading
      - port_of_discharge
...